Initial changes to get the app build and working
Forked and modified from the code of https://github.com/BaltiApps/Pixelify-Google-Photos related to tag v4.1

I'm not a developer and this code can be cleaned much more. I might do it once/if I learn how to do it and once/if I find time, for now it works for my purpose.